LAB-05/requirements.txt
requests==2.31.0
pillow==10.2.0
cryptography==42.0.2 